app/requirements.txt
flask==3.0.3
requests==2.31.0
beautifulsoup4==4.12.3
qbittorrent-api==2024.3.60
python-dotenv==1.0.1
transmission-rpc==7.0.11
deluge-web-client==1.3.0
cachetools==5.3.3
gunicorn==22.0.0